Fix for Solaris 32-bit size_t truncation from uint64_t

+++ plugin/command_log/command_log.cc
@@ -129,14 +129,14 @@ void CommandLog::apply(message::Command *to_apply)
    */
   string buffer= string(""); /* Buffer we will write serialized command to */
 
-  uint64_t length;
+  size_t length;
   ssize_t written;
   off_t cur_offset;
 
   to_apply->SerializeToString(&buffer);
 
   /* We force to uint64_t since this is what is reserved as the length header in the written log */
-  length= (uint64_t) buffer.length(); 
+  length= buffer.length(); 
 
   /*
    * Do an atomic increment on the offset of the log file position
